Moved testhelper.CreateBlock() to pkg/storage/tsdb/block
