+++ pkg/storegateway/bucket_e2e_test.go
@@ -36,7 +36,6 @@ import (
 	"github.com/grafana/mimir/pkg/storegateway/indexcache"
 	"github.com/grafana/mimir/pkg/storegateway/indexheader"
 	"github.com/grafana/mimir/pkg/storegateway/storepb"
-	"github.com/grafana/mimir/pkg/storegateway/testhelper"
 	"github.com/grafana/mimir/pkg/util"
 )
 
@@ -84,9 +83,9 @@ func prepareTestBlocks(t testing.TB, now time.Time, count int, dir string, bkt o
 
 		// Create two blocks per time slot. Only add 10 samples each so only one chunk
 		// gets created each. This way we can easily verify we got 10 chunks per series below.
-		id1, err := testhelper.CreateBlock(ctx, dir, series[:4], 10, mint, maxt, extLset)
+		id1, err := block.CreateBlock(ctx, dir, series[:4], 10, mint, maxt, extLset)
 		assert.NoError(t, err)
-		id2, err := testhelper.CreateBlock(ctx, dir, series[4:], 10, mint, maxt, extLset)
+		id2, err := block.CreateBlock(ctx, dir, series[4:], 10, mint, maxt, extLset)
 		assert.NoError(t, err)
 
 		dir1, dir2 := filepath.Join(dir, id1.String()), filepath.Join(dir, id2.String())
