+++ pkg/storegateway/indexheader/lazy_binary_reader_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/thanos-io/objstore/providers/filesystem"
 
 	"github.com/grafana/mimir/pkg/storage/tsdb/block"
-	"github.com/grafana/mimir/pkg/storegateway/testhelper"
 )
 
 func TestNewLazyBinaryReader_ShouldFailIfUnableToBuildIndexHeader(t *testing.T) {
@@ -47,7 +46,7 @@ func TestNewLazyBinaryReader_ShouldBuildIndexHeaderFromBucket(t *testing.T) {
 	t.Cleanup(func() { require.NoError(t, bkt.Close()) })
 
 	// Create block.
-	blockID, err := testhelper.CreateBlock(ctx, tmpDir, []labels.Labels{
+	blockID, err := block.CreateBlock(ctx, tmpDir, []labels.Labels{
 		labels.FromStrings("a", "1"),
 		labels.FromStrings("a", "2"),
 		labels.FromStrings("a", "3"),
@@ -90,7 +89,7 @@ func TestNewLazyBinaryReader_ShouldRebuildCorruptedIndexHeader(t *testing.T) {
 	t.Cleanup(func() { require.NoError(t, bkt.Close()) })
 
 	// Create block.
-	blockID, err := testhelper.CreateBlock(ctx, tmpDir, []labels.Labels{
+	blockID, err := block.CreateBlock(ctx, tmpDir, []labels.Labels{
 		labels.FromStrings("a", "1"),
 		labels.FromStrings("a", "2"),
 		labels.FromStrings("a", "3"),
@@ -132,7 +131,7 @@ func TestLazyBinaryReader_ShouldReopenOnUsageAfterClose(t *testing.T) {
 	t.Cleanup(func() { require.NoError(t, bkt.Close()) })
 
 	// Create block.
-	blockID, err := testhelper.CreateBlock(ctx, tmpDir, []labels.Labels{
+	blockID, err := block.CreateBlock(ctx, tmpDir, []labels.Labels{
 		labels.FromStrings("a", "1"),
 		labels.FromStrings("a", "2"),
 		labels.FromStrings("a", "3"),
@@ -182,7 +181,7 @@ func TestLazyBinaryReader_unload_ShouldReturnErrorIfNotIdle(t *testing.T) {
 	t.Cleanup(func() { require.NoError(t, bkt.Close()) })
 
 	// Create block.
-	blockID, err := testhelper.CreateBlock(ctx, tmpDir, []labels.Labels{
+	blockID, err := block.CreateBlock(ctx, tmpDir, []labels.Labels{
 		labels.FromStrings("a", "1"),
 		labels.FromStrings("a", "2"),
 		labels.FromStrings("a", "3"),
@@ -234,7 +233,7 @@ func TestLazyBinaryReader_LoadUnloadRaceCondition(t *testing.T) {
 	t.Cleanup(func() { require.NoError(t, bkt.Close()) })
 
 	// Create block.
-	blockID, err := testhelper.CreateBlock(ctx, tmpDir, []labels.Labels{
+	blockID, err := block.CreateBlock(ctx, tmpDir, []labels.Labels{
 		labels.FromStrings("a", "1"),
 		labels.FromStrings("a", "2"),
 		labels.FromStrings("a", "3"),
